test: move integration to happy path
